% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/air_lead_and_scaling_functions.R
\name{calculate_scaled_airPb}
\alias{calculate_scaled_airPb}
\title{Calculate temporally scaled air lead exposure estimates at specific locations.}
\usage{
calculate_scaled_airPb(locations)
}
\arguments{
\item{locations}{Data.frame with columns 'id', 'lat','lon', 'start_date', and 'end_date' at minimum.}
}
\value{
A numeric vector of air lead estimates (ug/m3).
}
\description{
\code{calculate_scaled_airPb()} is a wrapper function that estimates airborne lead exposures
at provided locations by calling \code{calculate_airPb()}, then temporally scales those
estimates using scaling factors computed by calling \code{calculate_scaling_factors()}.
This function is particularly useful for calculating exposures at the same locations on
different dates.
}
\examples{
my_data <- data.frame(id = rep(1,3),
    lat = c(39.19674, 39.19674,	39.19674),
    lon = c(-84.58260, -84.58260, -84.58260),
    start_date = c(as.Date("2010-01-08"), as.Date("2012-06-08"), as.Date("2015-04-09")),
    end_date = c(as.Date("2010-02-08"), as.Date("2012-07-08"), as.Date("2015-05-09")))

airPb_scaled <- calculate_scaled_airPb(my_data)
}
\references{
Cole Brokamp, Roman Jandarov, MB Rao, Grace LeMasters, Patrick Ryan. Exposure assessment
models for elemental components of particulate matter in an urban environment:
A comparison of regression and random forest approaches. Atmospheric Environment. 151. 1-11. 2017.
\url{http://dx.doi.org/10.1016/j.atmosenv.2016.11.066}
}
